refactor component menu actions inside control
